Move PlatformContextFactory
