+++ platform-sdk/swirlds-cli/build.gradle.kts
@@ -8,6 +8,7 @@ plugins {
 mainModuleInfo { runtimeOnly("org.hiero.consensus.pces.noop.impl.test.fixtures") }
 
 testModuleInfo {
+    requires("com.swirlds.common.test.fixtures")
     requires("com.swirlds.config.extensions.test.fixtures")
     requires("com.swirlds.platform.core.test.fixtures")
     requires("org.hiero.consensus.event.stream.test.fixtures")
